+++ src/argilla/server/contexts/datasets.py
@@ -31,6 +31,7 @@ from argilla.server.models import (
     ResponseStatus,
     ResponseValue,
     Suggestion,
+    VectorSettings,
 )
 from argilla.server.models.suggestions import SuggestionCreateWithRecordId
 from argilla.server.schemas.v1.datasets import DatasetCreate, FieldCreate, QuestionCreate, RecordsCreate
@@ -42,7 +43,7 @@ from argilla.server.security.model import User
 if TYPE_CHECKING:
     from sqlalchemy.ext.asyncio import AsyncSession
 
-    from argilla.server.schemas.v1.datasets import DatasetUpdate
+    from argilla.server.schemas.v1.datasets import DatasetUpdate, VectorSettingsCreate
     from argilla.server.schemas.v1.fields import FieldUpdate
     from argilla.server.schemas.v1.questions import QuestionUpdate
     from argilla.server.schemas.v1.suggestions import SuggestionCreate
@@ -51,7 +52,11 @@ LIST_RECORDS_LIMIT = 20
 
 
 async def get_dataset_by_id(
-    db: "AsyncSession", dataset_id: UUID, with_fields: bool = False, with_questions: bool = False
+    db: "AsyncSession",
+    dataset_id: UUID,
+    with_fields: bool = False,
+    with_questions: bool = False,
+    with_vectors_settings: bool = False,
 ) -> Dataset:
     query = select(Dataset).filter_by(id=dataset_id)
     options = []
@@ -59,6 +64,8 @@ async def get_dataset_by_id(
         options.append(selectinload(Dataset.fields))
     if with_questions:
         options.append(selectinload(Dataset.questions))
+    if with_vectors_settings:
+        options.append(selectinload(Dataset.vectors_settings))
     if options:
         query = query.options(*options)
     result = await db.execute(query)
@@ -208,6 +215,46 @@ async def delete_question(db: "AsyncSession", question: Question) -> Question:
         raise ValueError("Questions cannot be deleted for a published dataset")
 
     return await question.delete(db)
+
+
+async def get_vector_settings_by_id(db: "AsyncSession", vector_settings_id: UUID) -> Union[VectorSettings, None]:
+    result = await db.execute(
+        select(VectorSettings).filter_by(id=vector_settings_id).options(selectinload(VectorSettings.dataset))
+    )
+    return result.scalar_one_or_none()
+
+
+async def get_vector_settings_by_name_and_dataset_id(
+    db: "AsyncSession", name: str, dataset_id: UUID
+) -> Union[VectorSettings, None]:
+    return await VectorSettings.read_by(db, name=name, dataset_id=dataset_id)
+
+
+async def delete_vector_settings(db: "AsyncSession", vector_settings: VectorSettings) -> VectorSettings:
+    # TODO: for now the search engine does not allow to delete vector settings
+    return await vector_settings.delete(db)
+
+
+async def create_vector_settings(
+    db: "AsyncSession", search_engine: "SearchEngine", dataset: Dataset, vector_settings_create: "VectorSettingsCreate"
+) -> VectorSettings:
+    async with db.begin_nested():
+        vector_settings = await VectorSettings.create(
+            db,
+            name=vector_settings_create.name,
+            dimensions=vector_settings_create.dimensions,
+            description=vector_settings_create.description,
+            dataset_id=dataset.id,
+            autocommit=False,
+        )
+
+        if dataset.is_ready:
+            await db.flush([vector_settings])
+            await search_engine.configure_index_vectors(vector_settings)
+
+    await db.commit()
+
+    return vector_settings
 
 
 async def get_record_by_id(